Created OS specific header that will handle kernel event for specific OS. only including mac. Separated out example for standalone example and rest api example. Updated make to use some default make variables.
