+++ example/rest_api/assets/shiba_close.png
