+++ example/rest_api/generate_models.c
