+++ example/rest_api/main.c
@@ -345,7 +345,7 @@ int main()
   signal(SIGINT, handle_sigint);
 
   // Using Epoll Fd to assign server_fd and client_fd.
-  RunEpollLoop(server_fd);
+  RunEventLoop(server_fd);
 
   // TODO: Maybe assign fd into arena and clean up at the end easily.
   printf("Shutting down server...\n");
