+++ example/rest_api/paths/index.html
