+++ example/stand_alone/a.out
