+++ src/server.c
@@ -1,9 +1,7 @@
 #include <seobeo/helper.h>
 #include <seobeo/server.h>
+#include <seobeo/os.h>
 #include <sys/time.h>
-
-// Global epoll file descriptor - needed for proper cleanup
-static int global_epfd = -1;
 
 // --- Response -- //
 void GenerateResponseHeader
@@ -299,27 +297,6 @@ int MatchRoute(const char *pattern, const char *path, HttpRequestType *req)
   return *pattern_ptr == '\0' && *path_ptr == '\0';
 }
 
-void CleanupClient(int client_fd)
-{
-  // Disable Nagle's algorithm (optional for closed socket, but harmless)
-  int opt = 1;
-  setsockopt(client_fd, IPPROTO_TCP, TCP_NODELAY, &opt, sizeof(opt));
-
-  if (global_epfd != -1) {
-#if defined(__APPLE__) || defined(__FreeBSD__)
-    // kqueue: remove fd
-    struct kevent ev;
-    EV_SET(&ev, client_fd, EVFILT_READ, EV_DELETE, 0, 0, NULL);
-    kevent(global_epfd, &ev, 1, NULL, 0, NULL);
-#else
-    // epoll
-    epoll_ctl(global_epfd, EPOLL_CTL_DEL, client_fd, NULL);
-#endif
-  }
-
-  close(client_fd);
-}
-
 int send_all(int sockfd, const void *buf, size_t len)
 {
   const char *p = buf;
@@ -613,168 +590,4 @@ void WriteToLogs(const char *restrict format, ...)
     fprintf(file, "[%s] %s \n", timestamp, logger_buffer);
     fclose(file);
   }
-}
-
-void RunEpollLoop(const int server_fd)
-{
-  #if defined(__APPLE__) || defined(__FreeBSD__)
-  struct kevent change_event;
-  struct kevent events_buffer[MAX_EVENTS];
-
-  int kq = kqueue();
-  if (kq == -1) {
-    perror("kqueue");
-    return;
-  }
-
-  global_epfd = kq;
-
-  EV_SET(&change_event, server_fd, EVFILT_READ, EV_ADD, 0, 0, NULL);
-  if (kevent(kq, &change_event, 1, NULL, 0, NULL) == -1) {
-    perror("kevent register");
-    close(kq);
-    return;
-  }
-
-  while (!stop_server) {
-    int n = kevent(kq, NULL, 0, events_buffer, MAX_EVENTS, NULL);
-    if (n == -1) {
-      if (errno == EINTR) continue;
-      perror("kevent");
-      break;
-    }
-
-    for (int i = 0; i < n; i++) {
-      int fd = (int)events_buffer[i].ident;
-
-      if (fd == server_fd) {
-        while (1) {
-          int client_fd = accept(server_fd, NULL, NULL);
-          if (client_fd == -1) {
-            if (errno != EAGAIN && errno != EWOULDBLOCK) {
-              perror("accept");
-            }
-            break;
-          }
-
-          if (SetNonBlocking(client_fd) == -1) {
-            perror("SetNonBlocking");
-            close(client_fd);
-            continue;
-          }
-
-          int opt = 1;
-          setsockopt(client_fd, IPPROTO_TCP, TCP_NODELAY, &opt, sizeof(opt));
-
-          EV_SET(&change_event, client_fd, EVFILT_READ, EV_ADD | EV_CLEAR, 0, 0, NULL);
-          if (kevent(kq, &change_event, 1, NULL, 0, NULL) == -1) {
-            perror("kevent add client");
-            close(client_fd);
-            continue;
-          }
-        }
-      } else {
-        HandleRequest(fd);
-      }
-    }
-  }
-
-  close(kq);
-  global_epfd = -1;
-  #else
-  struct epoll_event events_buffer[MAX_EVENTS];
-  // EPOLLIN means ready for reading
-  struct epoll_event ev = {
-    .events = EPOLLIN,
-    .data.fd = server_fd
-  };
-  int epfd = epoll_create1(0);
-  if (epfd == -1)
-  {
-    perror("epoll_create1");
-    return;
-  }
-  
-  // Set global epfd for cleanup
-  global_epfd = epfd;
-
-  if (epoll_ctl(epfd, EPOLL_CTL_ADD, server_fd, &ev) == -1)
-  {
-    perror("epoll_ctl: server_fd");
-    close(epfd);
-    return;
-  }
-  
-  while (!stop_server)
-  {
-    // Find open events and overwrite to events 
-    int n = epoll_wait(epfd, events_buffer, MAX_EVENTS, 1000); // 1 second timeout
-    
-    if (n == -1)
-    {
-      if (errno == EINTR) continue; // Interrupted by signal
-      perror("epoll_wait");
-      break;
-    }
-  
-    // Look through all events that are waiting
-    for (int i = 0; i < n; i++)
-    {
-      int fd = events_buffer[i].data.fd;
-  
-      // If it is the server socket, it means a new client is trying to connect.
-      if (fd == server_fd)
-      {
-        while (1)
-        {
-          // Create a new client socket where client can bind to socket_fd
-          int client_fd = accept(server_fd, NULL, NULL);
-          // No more connections to accept to given server
-          if (client_fd == -1)
-          {
-            if (errno != EAGAIN && errno != EWOULDBLOCK)
-            {
-              perror("accept");
-            }
-            break;
-          }
-
-          if (SetNonBlocking(client_fd) == -1)
-          {
-            perror("SetNonBlocking");
-            close(client_fd);
-            continue;
-          }
-          
-          // Set TCP_NODELAY for new client connections
-          int opt = 1;
-          setsockopt(client_fd, IPPROTO_TCP, TCP_NODELAY, &opt, sizeof(opt));
-  
-          // Register this client socket with epoll.
-          // using EPOLLET so that it notifies when the fd is ready?
-          // not really sure about this one, but apparently more performant
-          struct epoll_event client_ev = {
-            .events = EPOLLIN | EPOLLET,  
-            .data.fd = client_fd
-          };
-          if (epoll_ctl(epfd, EPOLL_CTL_ADD, client_fd, &client_ev) == -1)
-          {
-            perror("epoll_ctl: client_fd");
-            close(client_fd);
-            continue;
-          }
-        }
-      }
-      else
-      {
-        // If it is client, then handle the request.
-        // Note: CleanupClient will remove from epoll and close the fd
-        HandleRequest(fd);
-      }
-    }
-  }
-
-  close(epfd);
-  global_epfd = -1;
-  #endif
 }
