--- stand_alone/paths/index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Simple Page</title>
-  <style>
-    body {
-      font-family: sans-serif;
-      background-color: #f4f4f4;
-      padding: 2em;
-    }
-    h1 {
-      color: #333;
-    }
-    p {
-      color: #555;
-    }
-  </style>
-</head>
-<body>
-  <h1>Hello, World!</h1>
-  <p>This is a simple HTML page.</p>
-</body>
-</html>
-
